+++ controllers/project_status_controller.js
@@ -15,6 +15,12 @@ export default class ProjectStatusController {
 
     // Create a project status
     async create(req, res) {
+
+        // Check if the status_name is provided
+        if (!req.body.status_name) {
+            return res.status(400).json({ error: 'Données manquantes' });
+        }
+
         try {
             const [results, fields] = await dbQuery('INSERT INTO project_status (status_name) VALUES (?)', [req.body.status_name]);
             res.json(results);
@@ -26,6 +32,12 @@ export default class ProjectStatusController {
 
     // Update a project status
     async update(req, res) {
+
+        // Check if the status_name and the id are provided
+        if (!req.body.status_name || !req.params.id) {
+            return res.status(400).json({ error: 'Données manquantes' });
+        }
+
         try {
             const [results, fields] = await dbQuery('UPDATE project_status SET status_name = ? WHERE id = ?', [req.body.status_name, req.params.id]);
             res.json(results);
@@ -37,6 +49,12 @@ export default class ProjectStatusController {
 
     // Delete a project status
     async delete(req, res) {
+
+        // Check if the id is provided
+        if (!req.params.id) {
+            return res.status(400).json({ error: 'Id manquant' });
+        }
+
         try {
             const [results, fields] = await dbQuery('DELETE FROM project_status WHERE id = ?', [req.params.id]);
             res.json(results);
