+++ controllers/tasks_controller.js
@@ -1,8 +1,15 @@
 import { dbQuery } from "../db.js";
 
 export default class TasksController {
+
     // Get all the tasks by project
     async listAll(req, res) {
+
+        // Check if the project_uuid is provided
+        if (!req.params.uuid) {
+            return res.status(400).json({ error: 'Uuid manquant' });
+        }
+
         try {
             const [results, fields] = await dbQuery(`SELECT tasks.id as task_id, task_name, task_description, task_status_id, status_name
             FROM tasks 
@@ -17,6 +24,12 @@ export default class TasksController {
 
     // Create a task
     async create(req, res) {
+
+        // Check if the task_name, the task_description and the project_uuid are provided
+        if (!req.body.task_name || !req.body.project_uuid || !req.body.task_description) {
+            return res.status(400).json({ error: 'Données manquantes' });
+        }
+
         try {
             const [results, fields] = await dbQuery('INSERT INTO tasks (task_name, task_description, project_uuid) VALUES (?, ?, ?)', [req.body.task_name, req.body.task_description, req.body.project_uuid]);
             res.json(results);
@@ -28,6 +41,13 @@ export default class TasksController {
 
     // Update a task
     async update(req, res) {
+
+        // Check if the task_name, the task_status_id, the task_description and the task id are provided
+        if (!req.body.task_name || !req.body.task_status_id || !req.body.task_description || !req.params.id) {
+            return res.status(400).json({ error: 'Données manquantes' });
+
+        }
+
         try {
             const [results, fields] = await dbQuery('UPDATE tasks SET task_name = ?, task_status_id = ?, task_description = ? WHERE id = ?', [req.body.task_name, req.body.task_status_id, req.body.task_description, req.params.id]);
             res.json(results);
@@ -39,6 +59,11 @@ export default class TasksController {
 
     // Delete a task
     async delete(req, res) {
+        // Check if the task id is provided
+        if (!req.params.id) {
+            return res.status(400).json({ error: 'Id manquant' });
+        }
+
         try {
             const [results, fields] = await dbQuery('DELETE FROM tasks WHERE id = ?', [req.params.id]);
             res.json(results);
